Debugged minor syntax issues and get_backup.
Couple minor issues with our python which we remedied, then we had to change the logic in pre_start to no longer use manta directly, replacing it with get_backup logic. Also replaced a few logging messages to be more representative of what was happening. Removed import of libmanta since it is no longer directly used here.