Initial class for WebDav backupper.
Added the dependencies for the 'webdavclient' library into the Dockerfile. Picked that library since it seemed the most used, but if there's a different, more preferred python library for interacting with WebDav, it shouldn't be a problem to switch it out. Currently, it's looking for a specific container I'm running, for the sake of testing, but this will obviously be changed asap.